change floats to doubles

+++ render/normal.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   normal.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ralba-ji <ralba-ji@student.42madrid.com    +#+  +:+       +#+        */
+/*   By: isastre- <isastre-@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/26 20:07:11 by ralba-ji          #+#    #+#             */
-/*   Updated: 2026/03/11 20:28:14 by ralba-ji         ###   ########.fr       */
+/*   Updated: 2026/03/31 17:44:13 by isastre-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 t_3dvector	get_normal(t_list *hit, t_3dvector hit_point)
 {
 	t_cylinder	*c;
-	float		m;
+	double		m;
 	t_3dvector	axis;
 	t_3dvector	normal;
 
@@ -27,9 +27,9 @@ t_3dvector	get_normal(t_list *hit, t_3dvector hit_point)
 	c = (t_cylinder *)hit->obj;
 	axis = vector_normalize(c->dir);
 	m = vec_dot(vector_sub(hit_point, c->pos), axis);
-	if (m <= 0.001f)
-		return (vector_scale(axis, -1.0f));
-	if (m >= (c->height - 0.001f))
+	if (m <= 0.001)
+		return (vector_scale(axis, -1.0));
+	if (m >= (c->height - 0.001))
 		return (axis);
 	normal = vector_sub(vector_sub(hit_point, c->pos), vector_scale(axis, m));
 	return (vector_normalize(normal));
